Parse canvas location create file canvas.ts
